Moving util to different package
