+++ src/main/java/org/jboss/aerogear/unifiedpush/DefaultPushSender.java
@@ -17,7 +17,7 @@
 package org.jboss.aerogear.unifiedpush;
 
 import net.iharder.Base64;
-import org.jboss.aerogear.unifiedpush.http.HttpRequestUtil;
+import org.jboss.aerogear.unifiedpush.utils.HttpRequestUtil;
 import org.jboss.aerogear.unifiedpush.message.MessageResponseCallback;
 import org.jboss.aerogear.unifiedpush.message.UnifiedMessage;
 import org.jboss.aerogear.unifiedpush.model.ProxyConfig;
