+++ src/main/java/org/jboss/aerogear/unifiedpush/utils/HttpRequestUtil.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.aerogear.unifiedpush.http;
+package org.jboss.aerogear.unifiedpush.utils;
 
 import java.io.IOException;
 import java.io.OutputStream;
